Broke out HashRouter class to be imported into main.js using browserify. #2
