+++ client/main.js
@@ -1,4 +1,5 @@
 const createElement = require('./utils/createElement.js')
+const HashRouter = require('./utils/hashRouter.js')
 
 function sendUrlPostRequest(urlSubmission) {
   fetch('/url-request', {
@@ -121,35 +122,6 @@ function transitionToTracklistForm(keyData) {
   const $oldTrackForm = document.querySelector('.track-form-1')
   if (!$oldTrackForm) {
     addTrackForm()
-  }
-}
-
-class HashRouter {
-  constructor($views) {
-    this.$views = $views
-    this.isListening = false
-  }
-  match(hash) {
-    if (hash === '') {
-      hash = '#url-form'
-    }
-    const hashComponents = hash.split('?')
-    const viewId = hashComponents[0].replace('#', '')
-    this.$views.forEach($view => {
-      if ($view.id === viewId) {
-        $view.classList.remove('hidden')
-      }
-      else {
-        $view.classList.add('hidden')
-      }
-    })
-  }
-  listen() {
-    if (this.isListening) return
-    window.addEventListener('hashchange', () => {
-      this.match(window.location.hash)
-    })
-    this.isListening = true
   }
 }
